fix: add tests to the belonging stack
