+++ cmd/kubectl-mongodb/migrate-to-mck/mongodb_test.go
@@ -32,6 +32,16 @@ func TestBuildMongodbOptions_FlagTranslation(t *testing.T) {
 	assert.Equal(t, "my-rs", opts.ResourceNameOverride)
 }
 
+func TestBuildMongodbOptions_InvalidMultiClusterNames(t *testing.T) {
+	ac := om.NewAutomationConfig(om.Deployment{
+		"processes":   []any{},
+		"replicaSets": []any{},
+	})
+
+	f := mongodbFlags{multiClusterNames: "  ,  ,  "}
+	_, err := buildMongodbOptions(context.Background(), nil, ac, &ProjectConfigs{}, nil, strings.NewReader(""), f)
+	assert.ErrorContains(t, err, "no valid cluster names")
+}
 
 func TestCollectPrometheusCreds_NoPrometheus(t *testing.T) {
 	ac := om.NewAutomationConfig(om.Deployment{
@@ -67,48 +77,7 @@ func TestCollectPrometheusCreds_EmptyPassword(t *testing.T) {
 	opts := &GenerateOptions{Namespace: "mongodb"}
 	scanner := bufio.NewScanner(strings.NewReader("\n"))
 	err := collectPrometheusCreds(context.Background(), nil, ac, opts, scanner, "")
-	assert.ErrorContains(t, err, "input cancelled")
-}
-
-func tlsEnabledAC() *om.AutomationConfig {
-	return om.NewAutomationConfig(om.Deployment{
-		"processes": []any{
-			map[string]any{
-				"name":        "host-0",
-				"processType": string(om.ProcessTypeMongod),
-				"args2_6": map[string]any{
-					"net": map[string]any{
-						"tls": map[string]any{"mode": "requireSSL"},
-					},
-				},
-			},
-		},
-		"replicaSets": []any{},
-	})
-}
-
-func TestEnsureTLS_InvalidFlagReturnsError(t *testing.T) {
-	ac := tlsEnabledAC()
-	opts := &GenerateOptions{}
-	err := ensureTLS(ac, opts, nil, "Invalid_Name!")
-	assert.ErrorContains(t, err, "not a valid Kubernetes resource name")
-}
-
-func TestEnsureTLS_InvalidInteractiveInputReprompts(t *testing.T) {
-	ac := tlsEnabledAC()
-	opts := &GenerateOptions{}
-	scanner := bufio.NewScanner(strings.NewReader("Invalid_Name!\nmdb\n"))
-	err := ensureTLS(ac, opts, scanner, "")
-	require.NoError(t, err)
-	assert.Equal(t, "mdb", opts.CertsSecretPrefix)
-}
-
-func TestEnsureTLS_ValidFlagUsedDirectly(t *testing.T) {
-	ac := tlsEnabledAC()
-	opts := &GenerateOptions{}
-	err := ensureTLS(ac, opts, nil, "mdb")
-	require.NoError(t, err)
-	assert.Equal(t, "mdb", opts.CertsSecretPrefix)
+	assert.ErrorContains(t, err, "cannot be empty")
 }
 
 func TestIsTLSEnabled_TLSMode(t *testing.T) {
@@ -130,9 +99,9 @@ func TestIsTLSEnabled_TLSMode(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			processMap := map[string]om.Process{
 				"host-0": {
-					"args2_6": map[string]any{
-						"net": map[string]any{
-							"tls": map[string]any{
+					"args2_6": map[string]interface{}{
+						"net": map[string]interface{}{
+							"tls": map[string]interface{}{
 								"mode": tt.mode,
 							},
 						},
@@ -147,9 +116,9 @@ func TestIsTLSEnabled_TLSMode(t *testing.T) {
 func TestIsTLSEnabled_SSLMode(t *testing.T) {
 	processMap := map[string]om.Process{
 		"host-0": {
-			"args2_6": map[string]any{
-				"net": map[string]any{
-					"ssl": map[string]any{
+			"args2_6": map[string]interface{}{
+				"net": map[string]interface{}{
+					"ssl": map[string]interface{}{
 						"mode": "requireSSL",
 					},
 				},
@@ -169,7 +138,7 @@ func TestIsTLSEnabled_NoArgs(t *testing.T) {
 func TestIsTLSEnabled_NoNet(t *testing.T) {
 	processMap := map[string]om.Process{
 		"host-0": {
-			"args2_6": map[string]any{},
+			"args2_6": map[string]interface{}{},
 		},
 	}
 	assert.False(t, isTLSEnabled(processMap))
