+++ cmd/kubectl-mongodb/migrate-to-mck/users_test.go
@@ -9,30 +9,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/mongodb/mongodb-kubernetes/controllers/om"
-	"github.com/mongodb/mongodb-kubernetes/controllers/operator/authentication"
-	kubernetesClient "github.com/mongodb/mongodb-kubernetes/mongodb-community-operator/pkg/kube/client"
 )
-
-func newFakeSecret(name, password string) *corev1.Secret {
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{Name: name, Namespace: "default"},
-		Data:       map[string][]byte{passwordSecretDataKey: []byte(password)},
-	}
-}
-
-func newFakeKubeClient(objects ...runtime.Object) kubernetesClient.Client {
-	scheme := runtime.NewScheme()
-	_ = corev1.AddToScheme(scheme)
-	clientObjs := make([]runtime.Object, len(objects))
-	copy(clientObjs, objects)
-	return kubernetesClient.NewClient(fake.NewClientBuilder().WithScheme(scheme).WithRuntimeObjects(clientObjs...).Build())
-}
 
 func writeTempFile(t *testing.T, content string) string {
 	t.Helper()
@@ -130,56 +109,20 @@ func TestGenerateUserCRs_ExistingSecrets_SkipsUnmappedUsers(t *testing.T) {
 	assert.Contains(t, y, "alice")
 }
 
-func TestCollectUserSecretNamesInteractively_ExplicitName(t *testing.T) {
+func TestCollectUserPasswords_SkipOnEnter(t *testing.T) {
 	ac := om.NewAutomationConfig(om.Deployment{
 		"processes":   []any{},
 		"replicaSets": []any{},
 	})
 	ac.Auth.AutoUser = "mms-automation"
-	alice := &om.MongoDBUser{Username: "alice", Database: "admin", Roles: []*om.Role{{Role: "read", Database: "test"}}}
-	ac.Auth.Users = []*om.MongoDBUser{alice}
-	_, err := authentication.ConfigureScramCredentials(alice, "correct-password", ac)
-	require.NoError(t, err)
+	ac.Auth.Users = []*om.MongoDBUser{
+		{Username: "alice", Database: "admin", Roles: []*om.Role{{Role: "read", Database: "test"}}},
+		{Username: "bob", Database: "admin", Roles: []*om.Role{{Role: "read", Database: "test"}}},
+	}
 
-	secret := newFakeSecret("my-alice-secret", "correct-password")
-	kubeClient := newFakeKubeClient(secret)
-	opts := &GenerateOptions{Namespace: "default"}
-	scanner := bufio.NewScanner(strings.NewReader("my-alice-secret\n"))
-	err = collectUserSecretNamesInteractively(t.Context(), kubeClient, ac, opts, scanner)
+	opts := &GenerateOptions{}
+	scanner := bufio.NewScanner(strings.NewReader("\n\n"))
+	err := collectUserPasswords(ac, opts, scanner)
 	require.NoError(t, err)
-	assert.Equal(t, "my-alice-secret", opts.ExistingUserSecrets["alice:admin"])
-}
-
-func TestCollectUserSecretNamesInteractively_DefaultName(t *testing.T) {
-	ac := om.NewAutomationConfig(om.Deployment{
-		"processes":   []any{},
-		"replicaSets": []any{},
-	})
-	ac.Auth.AutoUser = "mms-automation"
-	alice := &om.MongoDBUser{Username: "alice", Database: "admin", Roles: []*om.Role{{Role: "read", Database: "test"}}}
-	ac.Auth.Users = []*om.MongoDBUser{alice}
-	_, err := authentication.ConfigureScramCredentials(alice, "correct-password", ac)
-	require.NoError(t, err)
-
-	// suggested name is "alice-password"
-	secret := newFakeSecret("alice-password", "correct-password")
-	kubeClient := newFakeKubeClient(secret)
-	opts := &GenerateOptions{Namespace: "default"}
-	scanner := bufio.NewScanner(strings.NewReader("\n")) // press Enter to accept suggested name
-	err = collectUserSecretNamesInteractively(t.Context(), kubeClient, ac, opts, scanner)
-	require.NoError(t, err)
-	assert.Equal(t, "alice-password", opts.ExistingUserSecrets["alice:admin"])
-}
-
-func TestCollectUserSecretNamesInteractively_NoUsers(t *testing.T) {
-	ac := om.NewAutomationConfig(om.Deployment{
-		"processes":   []any{},
-		"replicaSets": []any{},
-	})
-	ac.Auth.AutoUser = "mms-automation"
-	opts := &GenerateOptions{Namespace: "default"}
-	scanner := bufio.NewScanner(strings.NewReader(""))
-	err := collectUserSecretNamesInteractively(t.Context(), nil, ac, opts, scanner)
-	require.NoError(t, err)
-	assert.Nil(t, opts.ExistingUserSecrets)
+	assert.Empty(t, opts.UserPasswords)
 }
